+++ src/std/ft_realloc.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_realloc.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: dpattij <dpattij@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/01/15 23:36:11 by dpattij           #+#    #+#             */
-/*   Updated: 2020/01/15 23:47:30 by dpattij          ###   ########.fr       */
+/*   Project: memeshell420                                ::::::::            */
+/*   Members: dpattij, tuperera                         :+:    :+:            */
+/*   Copyright: 2020                                   +:+                    */
+/*                                                    +#+                     */
+/*                                                   +#+                      */
+/*                                                  #+#    #+#                */
+/*   while (!(succeed = try()));                   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 
 char	*ft_realloc(
 		char *ptr,
-		int ocapacity,
-		int capacity)
+		size_t ocapacity,
+		size_t capacity)
 {
 	char	*new;
-	int		idx;
+	size_t	idx;
 
 	new = malloc(capacity * sizeof(char));
 	if (new == NULL)
